backend/stock_backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
apturl==0.5.2
asgiref==3.7.2
asn1crypto==1.4.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.6.3
Brlapi==0.8.3
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==5.0.3
djangorestframework==3.15.0
dnspython==2.6.1
duplicity==0.8.21
exceptiongroup==1.1.3
fastapi==0.103.2
fasteners==0.14.1
Flask==3.0.0
frozendict==2.4.0
future==0.18.2
gyp==0.1
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
idna==3.6
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
lockfile==0.12.2
louis==3.20.0
lxml==5.1.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.1.3
mechanize==0.4.7
monotonic==1.6
more-itertools==8.10.0
multitasking==0.0.11
mysql-connector-python==8.3.0
netifaces==0.11.0
numpy==1.26.2
oauthlib==3.2.0
olefile==0.46
pandas==2.1.3
paramiko==2.9.3
peewee==3.17.1
pexpect==4.8.0
pika==1.3.2
Pillow==9.0.1
ply==3.11
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.4.2
pydantic_core==2.10.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.6.2
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyTelegramBotAPI==4.15.4
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.1
pytz==2024.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.31.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
sshuttle==1.0.5
starlette==0.27.0
supervisor==4.2.1
systemd-python==234
typing_extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.0
usb-creator==0.3.7
uvicorn==0.23.2
vboxapi==1.0
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.0.0
wsproto==1.0.0
xdg==5
xkit==0.0.0
yfinance==0.2.37
zipp==1.0.0
